+++ src/Twig/Extension.php
@@ -2,6 +2,7 @@
 
 namespace App\Twig;
 
+use Locale;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Twig\Extension\AbstractExtension;
@@ -24,6 +25,7 @@ class Extension extends AbstractExtension
     {
         return [
             new TwigFunction('locale_url', [$this, 'getLocaleUrl']),
+            new TwigFunction('preferred_locale_url', [$this, 'getPreferredLocaleUrl']),
         ];
     }
 
@@ -36,5 +38,15 @@ class Extension extends AbstractExtension
         $params = array_merge($routeParams, ['_locale' => $locale]);
 
         return $this->urlGenerator->generate($route, $params);
+    }
+
+    public function getPreferredLocaleUrl(): string
+    {
+        $request = $this->requestStack->getMasterRequest();
+        $route = $request->attributes->get('_route');
+        $routeParams = $request->attributes->get('_route_params');
+        $routeParams['_locale'] = Locale::getPrimaryLanguage($request->getPreferredLanguage());
+
+        return $this->urlGenerator->generate($route, $routeParams);
     }
 }