+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Added**
     - support for Czech, English and Russian languages;
     - ability to switch between the available locales;
+    - 'Translate to your preferred language' link;
 - **Changed**
 - **Deprecated**
 - **Removed**